refactor org.apache.shardingsphere.sql.parser.api.visitor
