+++ shardingsphere-sql-parser/shardingsphere-sql-parser-sqlserver/src/main/java/org/apache/shardingsphere/sql/parser/visitor/SQLServerVisitorFacade.java
@@ -17,13 +17,13 @@
 
 package org.apache.shardingsphere.sql.parser.visitor;
 
-import org.apache.shardingsphere.sql.parser.api.visitor.DALVisitor;
-import org.apache.shardingsphere.sql.parser.api.visitor.DCLVisitor;
-import org.apache.shardingsphere.sql.parser.api.visitor.DDLVisitor;
-import org.apache.shardingsphere.sql.parser.api.visitor.DMLVisitor;
+import org.apache.shardingsphere.sql.parser.api.visitor.statement.DALVisitor;
+import org.apache.shardingsphere.sql.parser.api.visitor.statement.DCLVisitor;
+import org.apache.shardingsphere.sql.parser.api.visitor.statement.DDLVisitor;
+import org.apache.shardingsphere.sql.parser.api.visitor.statement.DMLVisitor;
 import org.apache.shardingsphere.sql.parser.api.visitor.SQLVisitorFacade;
-import org.apache.shardingsphere.sql.parser.api.visitor.RLVisitor;
-import org.apache.shardingsphere.sql.parser.api.visitor.TCLVisitor;
+import org.apache.shardingsphere.sql.parser.api.visitor.statement.RLVisitor;
+import org.apache.shardingsphere.sql.parser.api.visitor.statement.TCLVisitor;
 import org.apache.shardingsphere.sql.parser.visitor.impl.SQLServerDALVisitor;
 import org.apache.shardingsphere.sql.parser.visitor.impl.SQLServerDCLVisitor;
 import org.apache.shardingsphere.sql.parser.visitor.impl.SQLServerDDLVisitor;
